refactor(claude): separate CLI support concerns
